+++ packages/backend/src/domains/catchMind/useCases/catchMind.repository.port.ts
